improve FUSE read performance
Introduces a dict that maps file handles to nodes when opening files.
This eliminates the need to resolve the full file path when performing
an operation (e.g. read) on a file of which the handle is known.

Adds a check for reads where an offset/length is requested that exceed
the file size so as not to request a new stream chunk but reuse one
with that offset, if existent.