Operations which cause exceptions during a transaction will rollback.
During the mod_cursor operations, we could leave the operation with a
transaction that was always committed - even in the case where the
operation was incomplete due to an exception occurring. This is quite
counter-intuitive when using transactions - you expect a failure to
rollback any database changes when a problem occurs.

Replacing the commit in the mod_cursor class with a commit-or-rollback
based on whether the exception was raised should at least make the
use of the transactions consistent with what you might expect.